finish boat

+++ ooad-challenges-1/logistic-system/src/main/java/com/dsantos/Shipment.java
@@ -1,4 +1,6 @@
 package com.dsantos;
 
-public record Shipment(double volume, double weight, String destination) {
+import java.math.BigDecimal;
+
+public record Shipment(BigDecimal volume, BigDecimal weight, String destination) {
 }
